pyproject.toml
[project]
name = "AL1S"
version = "0.1.0"
description = "爱丽丝来夏莱当值日生啦"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "duckduckgo-search>=7.3.0",
    "httpx>=0.27.2",
    "numpy>=2.2.2",
    "openai>=1.61.0",
    "paho-mqtt>=2.1.0",
    "pvporcupine>=3.0.4",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "python-vlc>=3.0.21203",
    "pyttsx3>=2.98",
    "requests>=2.32.3",
    "speechrecognition[audio]>=3.14.1",
    "textual>=1.0.0",
]
